CoreXZ implementation
I have added CoreXZ functionality to the repetier firmware.  Essentially this is just a modification of the CoreXY code but integrated cleanly so that things are backwards compatible with existing CoreXY functionality.  The addition of this required adding adding drive systems 8 and 9 to the firmware.  This was done for a new printer from 3D Seed (www.3d-seed.com).  This is only the firmware for AVR, I have not altered any code for the DUE.  It has been tested and is running well.